+++ PandaHR.WebAPI/src/PandaHR.Api.Services/Contracts/ICVService.cs
@@ -16,7 +16,7 @@ namespace PandaHR.Api.Services.Contracts
         Task<IEnumerable<CVSummaryDTO>> GetUserCVsPreviewAsync(Guid userId, int? pageSize, int? page);
         Task<IEnumerable<CVforSearchDTO>> GetUserCVsAsync(Guid userId, int? pageSize = 10, int? page = 1);
         Task<IEnumerable<VacancySummaryDTO>> GetVacanciesForCV(Guid CVId, int? pageSize = 10, int? page = 1);
-        Task<IEnumerable<ISkillSetWithRatingModel<Guid>>> GetCVsByVacancy(Guid vacancyId, double threshold);
+        Task<IEnumerable<ISkillSetWithRatingModel<Guid>>> GetCVsByVacancy(Guid vacancyId, int threshold);
         Task AddSkillKnowledgeToCVAsync(SkillKnowledgeServiceModel model, Guid CVId);
         Task DeleteSkillKnowledgeFromCVAsync(Guid skillKnowledgeId);
         Task AddJobExperienceToCVAsync(JobExperienceServiceModel model, Guid CVId);
